Move all css to the specific Icons
