+++ src/components/job/BusyIcon.js
@@ -1,19 +1,34 @@
 import PropTypes from 'prop-types';
 import { Box } from '../publiq-ui/Box';
+import { getValueFromTheme } from '../publiq-ui/theme';
+
+const getValue = getValueFromTheme('jobStatusIcon');
 
 const BusyIcon = ({ className, ...props }) => {
   return (
     <Box
       className={className}
-      as="svg"
+      forwardedAs="svg"
       version="1.1"
       xmlns="http://www.w3.org/2000/svg"
       viewBox="10 30 20 20"
+      css="overflow: visible;"
       {...props}
     >
       <g className="busy" fillRule="evenodd" fill="none">
-        <circle id="busy-background" cx="20" cy="40" r="10" />
+        <circle
+          css={`
+            fill: ${getValue('backgroundColor')};
+          `}
+          id="busy-background"
+          cx="20"
+          cy="40"
+          r="10"
+        />
         <g
+          css={`
+            stroke: ${getValue('busy.spinnerStrokeColor')};
+          `}
           strokeWidth="2"
           transform="translate(14 34)"
           className="busy-spinner"
